+++ src/Workflow/OCAP.Workflow.Abstractions/Class1.cs
@@ -1,0 +1,6 @@
+﻿namespace OCAP.Workflow.Abstractions;
+
+public class Class1
+{
+
+}