+++ src/Workflow/OCAP.Workflow.Application/Class1.cs
@@ -1,0 +1,6 @@
+﻿namespace OCAP.Workflow.Application;
+
+public class Class1
+{
+
+}