+++ src/Workflow/OCAP.Workflow.Domain/Class1.cs
@@ -1,0 +1,6 @@
+﻿namespace OCAP.Workflow.Domain;
+
+public class Class1
+{
+
+}