+++ src/Workflow/OCAP.Workflow.Infrastructure/Class1.cs
@@ -1,0 +1,6 @@
+﻿namespace OCAP.Workflow.Infrastructure;
+
+public class Class1
+{
+
+}